Add files via upload
Complete Developer Documents found in one of the 1.5.x archives. 
I've been trying to figure out why the transclusion indictors are not working with any of the NG builds, and have been getting heavy into the code.  It doesn't seem like all that much has changed with respect to names, so this seems like a good find that should be part of the NG line.